requirements.txt
streamlit
pandas
python-docx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-api-keys
pytz